Make the player invincible for testing

+++ entity/Player.h
@@ -10,5 +10,5 @@ class Player : public EquippedEntity {
     int64_t x;
     int64_t y;
 
-    explicit Player(int64_t hp, int64_t x = 0, int64_t y = 0);
+    explicit Player(double hp, int64_t x = 0, int64_t y = 0);
 };
